initial grid setup/display
