+++ game-of-life/src/components/About.js
@@ -1,0 +1,2 @@
+//rules of game
+//about game
