+++ game-of-life/src/components/Controls.js
@@ -1,0 +1,7 @@
+//for control panel
+//start, stop, clear, pause?
+//generation # will also be here
+//randomize button? (addtoinal ftr)
+//speed input? (additional ftr)
+//manual steps? (additional ftr)
+//change grid zie? (additional ftr)
